Added menu fetch functionality to order page and handled errors
