+++ src/client/css-styles/styles-order.css
@@ -51,6 +51,23 @@ input.search-bar, textarea {
 
 }
 
+#refresh-container {
+    height: 50px;
+    width: 100%;
+    padding-top: 10px;
+}
+
+#refresh-button {
+    width: 60px;
+    color: white;
+    background-color: rgb(161, 51, 51);
+    float: left;
+    margin-left: 40px;
+}
+#refresh-button:hover {
+    background-color: rgb(148, 47, 47);
+}
+
 #menu {
     
     /* border: 2px solid black; */
