+++ src/client/menus.html
@@ -13,7 +13,6 @@
     <script src="//cdn.jsdelivr.net/npm/pouchdb@8.0.1/dist/pouchdb.min.js"></script>
     <script type="module" src="../js-script/script-menus.js" defer></script>
     <script type="module" src="../js-script/script-universal.js" defer></script>
-    <script type="module" src="../js-script/script-menu-management.js" defer></script>
 
 </head>
 <body>
@@ -90,7 +89,8 @@
 
                 <div id="menu">
                     <!--Menus are procedurally generated-->
-
+                    <h1>There Is No Available Menu For the Selected Criteria.</h1>
+                    <p>Please Choose a Different Location/Time or Refresh Menus.</p>
                 </div>
 
             </div>
