+++ src/client/order.html
@@ -53,41 +53,17 @@
                     <input type="text" class="search-bar" placeholder="Enter Food Item (or click below)">
                     <button id="next-button" type="button">Next</button>
                 </div>
+                <div id="refresh-container">
+                    <button id="refresh-button">
+                        Refresh Menus
+                    </button>
+                </div>
 
                 <div id="menu">
 
-
-                    <!--These lists will be procedurally generated as we fetch from the database-->
-                    <div class="food-type-container">
-
-                        <div class="food-type-title">
-                            Entree (test)
-                        </div>
-                        <div class="food-option-list">
-                            <div class="food-option">Chicken</div>
-                            <div class="food-option">Fries</div>
-                            <div class="food-option">Burger</div>
-                            <div class="food-option">Grilled Cheese</div>
-                            <div class="food-option">Pizza</div>
-                            <div class="food-option">Pasta</div>
-                            <div class="food-option">Crab Rangoons</div>
-                            <div class="food-option">Soup</div>
-                            <div class="food-option">Burrito</div>
-                            <div class="food-option">Pork</div>
-                            <div class="food-option">Ribs</div>
-                            <div class="food-option">Seafood</div>
-                            <div class="food-option">Biscuit</div>
-                            <div class="food-option">Fried Chicken</div>
-                            <div class="food-option">Dumplings</div>
-                            <div class="food-option">Stir Fry</div>
-                            <div class="food-option">Sushi</div>
-                            <div class="food-option">Chicken Curry</div>
-                            <div class="food-option">Naan</div>
-                            <div class="food-option">Vegetables</div>
-                            <div class="food-option">Pho</div>
-                        </div>
-
-                    </div>
+                    <h1>There Is No Available Menu For the Selected Criteria.</h1>
+                    <p>Please Choose a Different Location or Refresh Menus.</p>
+                    <!--Lists will be procedurally generated as we fetch from the database-->
 
                 </div>
